+++ uliketu-service-api/src/main/java/com/hengsu/uliketu/core/service/WebsiteService.java
@@ -17,6 +17,6 @@ public interface WebsiteService{
 	
 	public int deleteByPrimaryKey(Long id);
 	
-	public int selectCount(WebsiteModel websiteModel);
+	public long selectCount(WebsiteModel websiteModel);
 	
 }
