+++ uliketu-webapp/src/main/java/com/hengsu/uliketu/core/vo/CashVO.java
@@ -8,7 +8,7 @@ public class CashVO{
 	
 	private Long id;
 	private Long userid;
-	private Double balance;
+	private Long balance;
 	private Double money;
 	private Double rate;
 	private Double poundage;
@@ -32,11 +32,11 @@ public class CashVO{
 		return this.userid;
 	}
 		
-	public void setBalance(Double balance){
+	public void setBalance(Long balance){
 		this.balance = balance;
 	}
 	
-	public Double getBalance(){
+	public Long getBalance(){
 		return this.balance;
 	}
